Move the tests into the existing wasm folder, and share the suite with the others.
