+++ tests/integration/telemetry/stats/prometheus/wasm/testdata/registry-secret.yaml
