+++ tests/integration/telemetry/stats/prometheus/wasm/testdata/wasm-filter.yaml
