+++ tests/integration/telemetry/stats/prometheus/wasm/wasm_modules/header_injector/src/lib.rs
